build: fix missing dist-info dir in installer
The installer currently doesn't include the dist-info directory of the
`streamlink` package. This causes issues with `importlib.metadata` when
it tries to find the package's metadata.

This results in the following error when `--loglevel=debug` is set,
as it lists the versions of Streamlink and its dependencies:
> importlib.metadata.PackageNotFoundError:
> No package metadata was found for streamlink

While Streamlink's dependencies are defined as wheels in pynsist's
build config and thus include their own dist-info directories that are
then copied to the pkgs directory of the installer, the `streamlink`
and `streamlink_cli` package names have to be explicitly defined, which
means that Streamlink's dist-info directory won't be copied by pynsist.

The only workaround for this seems to be adding the dist-info directory
to the list of custom files of the installer.

Changes:

- Don't install streamlink in current venv and use a temp dir instead,
  similar to the portable build config
- Remove unneeded and non-deterministic metadata from the dist-info dir
- Set `PYTHONPATH` for loading streamlink from the temp dir
- Add dist-info dir to custom files list in the installer build config